+++ packages/terra-dialog-modal/src/terra-dev-site/test/dialog-modal/DialogModalWithSelect.test.jsx
@@ -5,7 +5,7 @@ import ActionFooter from 'terra-action-footer';
 import Select from 'terra-form-select';
 import classNames from 'classnames/bind';
 import DialogModal from '../../../DialogModal';
-import styles from './DiaglogModalWithSelect.test.module.scss';
+import styles from './DialogModalWithSelect.test.module.scss';
 
 const cx = classNames.bind(styles);
 
@@ -49,7 +49,7 @@ class DefaultDialogModal extends React.Component {
           footer={<ActionFooter start="Footer Goes here" />}
         >
           <p>{text}</p>
-          <Select placeholder="Select a color" className={cx('selection-list')}>
+          <Select variant="combobox" placeholder="Select a color" className={cx('selection-list')}>
             <Select.Option value="blue" display="Blue" />
             <Select.Option value="green" display="Green" />
             <Select.Option value="purple" display="Purple" />