+++ packages/terra-date-input/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Changed
+  * Updated to support a peer dependency of react-intl v2-v5
+  * Updated `DateInputField` to use just `injectIntl` instead of `FormattedMessage` for react-intl v2-v5 compatibility
+
 ## 1.21.0 - (October 13, 2020)
 
 * Changed